Refactor generic schema analysis
