+++ iceaxe/schemas/cli.py
@@ -1,0 +1,2 @@
+async def create_all():
+    pass
